raffinerie-iot/requirements.txt
paho-mqtt==1.6.1
kafka-python==2.0.2
pyspark==3.5.0
pandas==2.0.3
numpy==1.24.3
